codes/task5/requirements.txt
torch>=1.10
torchvision
pandas
scikit-learn
matplotlib
seaborn
tqdm
Pillow
icrawler
lxml
opencv-python
grad-cam